+++ include/ftxui/component/screen_interactive.hpp
@@ -20,20 +20,28 @@ using Component = std::shared_ptr<ComponentBase>;
 
 class ScreenInteractive : public Screen {
  public:
+  using Callback = std::function<void()>;
+
   static ScreenInteractive FixedSize(int dimx, int dimy);
   static ScreenInteractive Fullscreen();
   static ScreenInteractive FitComponent();
   static ScreenInteractive TerminalOutput();
 
   void Loop(Component);
-  std::function<void()> ExitLoopClosure();
+  Callback ExitLoopClosure();
 
   void PostEvent(Event event);
   CapturedMouse CaptureMouse();
 
+  // Decorate a function. The outputted one will execute similarly to the
+  // inputted one, but with the currently active screen terminal hooks
+  // temporarily uninstalled.
+  Callback WithRestoredIO(Callback);
+
  private:
   void Install();
   void Uninstall();
+
   void Main(Component component);
   ScreenInteractive* suspended_screen_ = nullptr;